refactor: rename function format-hashtags.js and update imports
